requirements.txt
# Основная библиотека для работы с Tinkoff Invest API
tinkoff-investments>=0.2.0b60

# Работа с конфигурацией
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0
python-dotenv>=1.0.0

# База данных
SQLAlchemy>=2.0.0
aiosqlite>=0.19.0

# Логирование
loguru>=0.7.0

# Telegram-бот для уведомлений
python-telegram-bot>=20.0

# Утилиты
aiohttp>=3.8.0
async-timeout>=4.0.0
